Updated two methods in the DocumentService class and added missing functionality.
Currently the is no functionality in the new DigiDoc4j library that would allow to test whether we are dealing with a test card.
The previous library JDigiDoc had this, so I have taken the related parts from there.
Should the required functionality appear in the future versions of DigiDoc4j we can utilize that instead.