Fixed Spring configurations. Now @Transactional works as it should.
Previous Spring configuration was wrong in the meaning that it SILENTLY prevented @Transactional functionality.
The reason was that <tx:annotation-driven transaction-manager="transactionManager" /> was configured in the application context,
while beans that have @Transactional methods was configured in the related servlet context.

More information and details can be found in the following  answer on Stack Overflow:
http://stackoverflow.com/a/10021874/814702